Responsive Update Part 1 - Commit 5
Made it so plus ones do not reload, same with comments, and deleting posts. Plus, ones do not work on communities because that would be some work. I think, and I am lazy.

Redesigned how commenting looks to be me accurate. 

Made it so editing communities do not reload, same with uploading images on profiles.